requirements.txt
llama-cpp-python==0.2.11
discord.py==2.3.2
python-dotenv==1.0.0
